# optimize cte distinct joins