--- a/models/silver/liquidity_pool/raydium/clmm/silver__liquidity_pool_actions_raydium_clmm.sql
+++ b/models/silver/liquidity_pool/raydium/clmm/silver__liquidity_pool_actions_raydium_clmm.sql
@@ -120,13 +120,17 @@ WITH base AS (
 
 transfers AS (
     SELECT 
-        * exclude(index),
-        split_part(index,'.',1)::int AS index,
-        nullif(split_part(index,'.',2),'')::int AS inner_index
+        t.* exclude(index),
+        split_part(t.index,'.',1)::int AS index,
+        nullif(split_part(t.index,'.',2),'')::int AS inner_index
     FROM
-        {{ ref('silver__transfers') }}
+        {{ ref('silver__transfers') }} AS t
+    INNER JOIN
+        (SELECT DISTINCT block_timestamp::date AS bt, tx_id FROM base) AS b
+        ON b.bt = t.block_timestamp::date
+        AND b.tx_id = t.tx_id
     WHERE
-        succeeded
+        t.succeeded
         AND {{ between_stmts }}
 ),