--- a/models/silver/nfts/silver__nft_sales_magic_eden_v2_decoded.sql
+++ b/models/silver/nfts/silver__nft_sales_magic_eden_v2_decoded.sql
@@ -95,10 +95,10 @@ transfers AS (
         a._inserted_timestamp 
     FROM
         {{ ref('silver__transfers') }} a
-    INNER JOIN (
-        SELECT DISTINCT tx_id
-        FROM decoded
-    ) d ON d.tx_id = a.tx_id
+    INNER JOIN 
+        (SELECT DISTINCT tx_id, block_timestamp::date as bt FROM decoded) d 
+        ON d.tx_id = a.tx_id
+        AND d.bt = a.block_timestamp::date
     WHERE
         a.succeeded
         AND {{ between_stmts }}