--- a/models/silver/protocols/marinade/silver__marinade_liquid_staking_actions.sql
+++ b/models/silver/protocols/marinade/silver__marinade_liquid_staking_actions.sql
@@ -57,14 +57,9 @@ mints AS (
         a.*
     FROM {{ ref('silver__token_mint_actions') }} a
         INNER JOIN (
-            SELECT 
-                DISTINCT tx_id
-            FROM 
-                base
-            WHERE 
-                event_type IN ('deposit', 'depositStakeAccount')
-        ) b 
-        ON b.tx_id = a.tx_id
+            SELECT DISTINCT tx_id, block_timestamp::date as bt FROM base WHERE event_type IN ('deposit', 'depositStakeAccount')) b 
+        ON b.tx_id = a.tx_id 
+        AND b.bt = a.block_timestamp::date
     WHERE
         a.succeeded
         AND {{ between_stmts }}
@@ -78,13 +73,13 @@ transfers AS (
         {{ ref('silver__transfers') }} a
         INNER JOIN (
             SELECT 
-                DISTINCT tx_id
+                DISTINCT tx_id, block_timestamp::date as bt
             FROM 
                 base
             WHERE 
                 event_type = 'claim'
         ) b 
-        ON b.tx_id = a.tx_id
+        ON b.tx_id = a.tx_id and b.bt = a.block_timestamp::date
     WHERE
         a.succeeded
         AND {{ between_stmts }}
@@ -96,13 +91,13 @@ sol_balances as (
         {{ ref('silver__sol_balances') }} a
         INNER JOIN (
             SELECT 
-                DISTINCT tx_id
+                DISTINCT tx_id, block_timestamp::date as bt
             FROM 
                 base
             WHERE 
                 event_type = 'depositStakeAccount'
         ) b 
-        ON b.tx_id = a.tx_id
+        ON b.tx_id = a.tx_id and b.bt = a.block_timestamp::date
     WHERE
         a.succeeded
         AND {{ between_stmts }}