Optimize joins by filtering records in tables such as `transfers` with inner joins to base tables
- this initial PR makes updates to the tables that use `decoded_instructions_combined` and `{{ between_stmts }}` logic. Other models will be updated in future PRs.

Results in significant run time improvements, ie:
old:
-  `20:05:34  1 of 1 OK created sql incremental model silver.liquidity_pool_actions_raydium_clmm  [SUCCESS 1921 in 105.12s]`
- `20:30:55  1 of 1 OK created sql incremental model silver.liquidity_pool_actions_orca_whirlpool  [SUCCESS 38740 in 253.82s]`

new:
- `20:01:40  1 of 1 OK created sql incremental model silver.liquidity_pool_actions_raydium_clmm  [SUCCESS 1921 in 26.61s]`
- `20:33:14  1 of 1 OK created sql incremental model silver.liquidity_pool_actions_orca_whirlpool  [SUCCESS 38740 in 67.20s]`

- Runs on medium:
```
21:34:41  4 of 12 OK created sql incremental model silver.liquidity_pool_actions_orcav2 .. [SUCCESS 1 in 21.97s]
21:34:42  1 of 12 OK created sql incremental model silver.bridge_mayan_transfers_decoded . [SUCCESS 817 in 22.96s]
21:34:43  3 of 12 OK created sql incremental model silver.liquidity_pool_actions_orcav1 .. [SUCCESS 2 in 24.81s]
21:34:44  2 of 12 OK created sql incremental model silver.liquidity_pool_actions_orca_whirlpool  [SUCCESS 29742 in 25.24s]
21:34:56  5 of 12 OK created sql incremental model silver.liquidity_pool_actions_raydium_clmm  [SUCCESS 7 in 15.18s]
21:34:56  6 of 12 OK created sql incremental model silver.liquidity_pool_actions_raydium_cpmm  [SUCCESS 1 in 14.46s]
21:34:59  8 of 12 OK created sql incremental model silver.marinade_liquid_staking_actions  [SUCCESS 23 in 14.99s]
21:34:59  7 of 12 OK created sql incremental model silver.liquidity_pool_actions_raydiumv4  [SUCCESS 36 in 15.87s]
21:35:08  9 of 12 OK created sql incremental model silver.nft_sales_hadeswap_decoded ..... [SUCCESS 2 in 12.70s]
21:35:12  11 of 12 OK created sql incremental model silver.nft_sales_solsniper ........... [SUCCESS 1 in 13.06s]
21:35:13  12 of 12 OK created sql incremental model silver.nft_sales_tensorswap .......... [SUCCESS 75 in 13.20s]
21:35:13  10 of 12 OK created sql incremental model silver.nft_sales_magic_eden_v2_decoded  [SUCCESS 72 in 16.46s]
22:11:59  1 of 1 OK created sql incremental model silver.swaps_intermediate_phoenix ...... [SUCCESS 8185 in 46.90s]
```